renamed 'demo' user to 'lack', generate_recipe.sh works with directories
hacked generate_recipe.sh to work with a directory instead of a debian package
name; also included the ability to change the UID/GID, and accepting a regular
expression so that you can rename where files go on the target.  renamed the
'demo' user to 'lack', which included changing all of the filelists with
hardcoded paths and the entries in group/shadow/passwd.  Added the Rojo mods
to the Nodoka GTK theme.